fix: use `double` for Hz values
Fixes a compiler warning about implicit type conversion.

+++ src/track/track.cpp
@@ -1775,8 +1775,8 @@ ExportTrackMetadataResult Track::exportMetadata(
         normalizedFromRecord.normalizeBeforeExport();
         // Encode tuning offset (RapidEvolution style) into key text for tag roundtrip.
         // Keep the database value untouched; only the exported metadata is modified.
-        const int tuningHz = m_record.getKeys().getTuningFrequencyHz();
-        if (tuningHz > 0 && tuningHz != 440) {
+        const double tuningHz = m_record.getKeys().getTuningFrequencyHz();
+        if (tuningHz > 0.0 && tuningHz != 440.0) {
             QString keyText = normalizedFromRecord.getTrackInfo().getKeyText();
             if (keyText.isEmpty()) {
                 const auto key = m_record.getKeys().getGlobalKey();
@@ -1785,7 +1785,7 @@ ExportTrackMetadataResult Track::exportMetadata(
                 }
             }
             if (!keyText.isEmpty()) {
-                const double cents = 1200.0 * std::log2(static_cast<double>(tuningHz) / 440.0);
+                const double cents = 1200.0 * std::log2(tuningHz / 440.0);
                 const int centsRounded = static_cast<int>(std::lround(cents));
                 if (centsRounded != 0) {
                     const QString offsetText = centsRounded > 0
